Replace ConstantDiscovery and remove ConstantResolver
