+++ lib/packwerk/run_context.rb
@@ -15,7 +15,6 @@ module Packwerk
       def from_configuration(configuration)
         new(
           root_path: configuration.root_path,
-          load_paths: configuration.load_paths,
           package_paths: configuration.package_paths,
           inflector: ActiveSupport::Inflector,
           custom_associations: configuration.custom_associations,
@@ -23,6 +22,7 @@ module Packwerk
           cache_enabled: configuration.cache_enabled?,
           cache_directory: configuration.cache_directory,
           config_path: configuration.config_path,
+          loaders: configuration.loaders
         )
       end
     end
@@ -30,7 +30,6 @@ module Packwerk
     sig do
       params(
         root_path: String,
-        load_paths: T::Hash[String, Module],
         inflector: T.class_of(ActiveSupport::Inflector),
         cache_directory: Pathname,
         config_path: T.nilable(String),
@@ -39,11 +38,11 @@ module Packwerk
         associations_exclude: T::Array[String],
         checkers: T::Array[Checker],
         cache_enabled: T::Boolean,
+        loaders: T::Enumerable[Zeitwerk::Loader]
       ).void
     end
     def initialize(
       root_path:,
-      load_paths:,
       inflector:,
       cache_directory:,
       config_path: nil,
@@ -51,10 +50,11 @@ module Packwerk
       custom_associations: [],
       associations_exclude: [],
       checkers: Checker.all,
-      cache_enabled: false
+      cache_enabled: false,
+      loaders: []
     )
       @root_path = root_path
-      @load_paths = load_paths
+      @loaders = loaders
       @package_paths = package_paths
       @inflector = inflector
       @custom_associations = custom_associations
@@ -109,11 +109,10 @@ module Packwerk
 
     sig { returns(ConstantDiscovery) }
     def context_provider
-      @context_provider ||= ConstantDiscovery.for(
+      @context_provider ||= ConstantDiscovery.new(
         package_set,
         root_path: @root_path,
-        load_paths: @load_paths,
-        inflector: @inflector
+        loaders:   @loaders
       )
     end
