+++ test/fixtures/blank/config/environment.rb
@@ -1,0 +1,1 @@
+# this file intentionally left blank
