+++ test/unit/packwerk/cli_test.rb
@@ -7,7 +7,7 @@ require "support/rails_test_helper"
 module Packwerk
   class CliTest < Minitest::Test
     include TypedMock
-    include ApplicationFixtureHelper
+    include RailsApplicationFixtureHelper
 
     setup do
       setup_application_fixture
@@ -27,7 +27,6 @@ module Packwerk
       offense = Offense.new(file: file_path, message: violation_message)
 
       configuration = Configuration.new({ "parallel" => false })
-      configuration.stubs(load_paths: {})
       RunContext.any_instance.stubs(:process_file).at_least_once.returns([offense])
 
       string_io = StringIO.new
@@ -56,7 +55,6 @@ module Packwerk
       offense = Offense.new(file: file_path, message: violation_message)
 
       configuration = Configuration.new({ "parallel" => false })
-      configuration.stubs(load_paths: {})
 
       RunContext.any_instance.stubs(:process_file)
         .at_least(2)
@@ -183,7 +181,6 @@ module Packwerk
       offense = Offense.new(file: file_path, message: violation_message)
 
       configuration = Configuration.new
-      configuration.stubs(load_paths: {})
       RunContext.any_instance.stubs(:process_file)
         .returns([offense])
